% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquaint-package.R
\docType{package}
\name{acquaint-package}
\alias{acquaint}
\alias{acquaint-package}
\title{acquaint: Model Context Protocol For Your R Session}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of acquaint is to enable LLM-enabled tools like Claude Code to learn about the R packages you have installed using the [Model Context Protocol](https://modelcontextprotocol.io/) (MCP). Tools configured with acquaint can peruse package documentation and learn to use R packages even if they weren't included in their training data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/acquaint}
  \item \url{https://simonpcouch.github.io/acquaint/}
  \item Report bugs at \url{https://github.com/simonpcouch/acquaint/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Authors:
\itemize{
  \item Winston Chang \email{winston@posit.co}
  \item Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
