% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R, R/proxy.R
\name{mcp}
\alias{mcp}
\alias{mcp_host}
\alias{mcp_proxy}
\title{Model context protocol for your R session}
\usage{
mcp_host()

mcp_proxy()
}
\description{
Together, these functions implement a model context protocol server for your
R session.
}
\section{Configuration}{


\code{\link[=mcp_proxy]{mcp_proxy()}} should be configured with the MCP clients via the \code{Rscript}
command. For example, to use with Claude Desktop, paste the following in your
Claude Desktop configuration (on macOS, at
\code{file.edit("~/Library/Application Support/Claude/claude_desktop_config.json")}):

\if{html}{\out{<div class="sourceCode json">}}\preformatted{\{
  "mcpServers": \{
    "r-acquaint": \{
      "command": "Rscript",
      "args": ["-e", "acquaint::mcp_proxy()"]
    \}
  \}
\}
}\if{html}{\out{</div>}}

Or, to use with Claude Code, you might type in a terminal:

\if{html}{\out{<div class="sourceCode bash">}}\preformatted{claude mcp add -s "user" r-acquaint Rscript -e "acquaint::mcp_proxy()"
}\if{html}{\out{</div>}}

\strong{mcp_proxy() is not intended for interactive use.}

The proxy interfaces with the MCP client on behalf of the host in
your R session. \strong{Use \code{\link[=mcp_host]{mcp_host()}} to start the host in your R session.}
Place a call to \code{acquaint::mcp_host()} in your \code{.Rprofile}, perhaps with
\code{usethis::edit_r_profile()}, to start a host for every interactive R session
you start.
}

\examples{
if (interactive()) {
mcp_host()
}

}
