% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp-config.R
\name{mcp_config}
\alias{mcp_config}
\title{Generate model context protocol configuration}
\usage{
mcp_config()
}
\value{
A character vector containing the MCP configuration JSON, invisibly. The
function also prints the configuration to the console.
}
\description{
This function provides the \code{.json} needed to configure rmcp with LLM-enabled
tools like Claude Code or Claude for Desktop.
}
\examples{
mcp_config()

}
