% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquaint-package.R
\docType{package}
\name{acquaint-package}
\alias{acquaint}
\alias{acquaint-package}
\title{acquaint: Model Context Protocol Server For Your R Sessions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements a model context protocol (MCP) server for your R sessions, allowing MCP-compatible apps like Claude Desktop and Claude Code to run R code in your active R sessions in order to better answer questions. By default, the package supplies a set of tools that allow models to inspect your R environment and peruse package documentation, but supports arbitrary R functions as tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/acquaint}
  \item \url{https://simonpcouch.github.io/acquaint/}
  \item Report bugs at \url{https://github.com/simonpcouch/acquaint/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Authors:
\itemize{
  \item Winston Chang \email{winston@posit.co}
  \item Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
